refactor: [#31] move test
